backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.23
pydantic==2.5.3
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
# pymysql==1.1.0  # Not needed for SQLite (SQLite is built into Python)
